Change mkdir method (makes parents now)
The mkdir function was using os.mkdir, which wasn't creating parent directories, and the script was therefore failing. Rather than explicitly creating each parent, it seems a fair use to change the os.mkdir function to os.makedirs (which does automatically create parents). This allows the script to finish successfully.