Return absolute path of the downloaded file and use this path for extraction
Modified the `utils.download_remote_file` function in `utils.py` to return the absolute path of the downloaded file instead of relative path. This change extends the robustness as it makes sure we are always unzipping the downloaded file even if cwd changes. Also, adjusted the `automx.py` script to use the absolute path returned from the download function when extracting the file. This ensures that the correct file gets extracted regardless of current working directory.