Modify wget command to specify output file
In utils.py, the wget command was updated in the exec_cmd function. Previously, the command did not specify the output file for the downloaded content. Now, a new target parameter has been added to the wget command to direct the download to a specific file. This change is done to provide more control over the file handling process within the function.